Add PSC example
