+++ docs/schema.yml
@@ -105,6 +105,7 @@ properties:
                           sigma: {type: number, default: 0.0, description: "Diameter (Å)"}
                           tension: {type: number, default: 0.0, description: "Surface tension (kJ/mol/Å²)"}
                           tfe: {type: number, default: 0.0, description: "Transfer free energy (kJ/mol/Å²/M)"}
+                          psc: {type: object, description: "Patchy spherocylinder properties"}
                 - type: string
 
     moleculelist:
