+++ src/CMakeLists.txt
@@ -64,7 +64,7 @@ set(objs analysis.cpp average.cpp atomdata.cpp auxiliary.cpp bonds.cpp celllisti
 
 set(hdrs analysis.h average.h atomdata.h auxiliary.h bonds.h celllist.h celllistimpl.h
 	chainmove.h clustermove.h core.h forcemove.h energy.h externalpotential.h geometry.h group.h io.h
-	molecule.h montecarlo.h move.h mpicontroller.h particle.h penalty.h potentials.h
+	molecule.h montecarlo.h move.h mpicontroller.h particle.h penalty.h potentials_base.h potentials.h
 	reactioncoordinate.h rotate.h sasa.h smart_montecarlo.h space.h speciation.h spherocylinder.h
         random.h regions.h tensor.h units.h aux/arange.h
 	aux/eigen_cerealisation.h aux/eigensupport.h aux/iteratorsupport.h aux/multimatrix.h
